rename elements POM folder
